requirements.txt
ultralytics>=8.1.0
supervision>=0.21.0
deep-sort-realtime>=1.3.2
numpy>=1.23
opencv-python>=4.8.0
scipy>=1.10
torch>=2.1.0
torchvision>=0.16.0
PyYAML>=6.0.1
